Add optional null check logic in find query.

+++ user-service/user-service/src/main/java/intelligent_pay/userservice/command/MemberCommandService.java
@@ -1,15 +1,15 @@
 package intelligent_pay.userservice.command;
 
+import intelligent_pay.userservice.controller.restResponse.ResponseMessage;
 import intelligent_pay.userservice.domain.Member;
 import intelligent_pay.userservice.dto.changeInfo.ChangeEmailRequest;
 import intelligent_pay.userservice.dto.changeInfo.ChangePasswordRequest;
-import intelligent_pay.userservice.dto.changeInfo.WithdrawRequest;
 import intelligent_pay.userservice.dto.signupAndLogin.MemberLoginRequest;
 import intelligent_pay.userservice.dto.signupAndLogin.MemberSignupRequest;
+import intelligent_pay.userservice.exception.MemberCustomException;
 import intelligent_pay.userservice.jwt.JwtTokenProvider;
 import intelligent_pay.userservice.jwt.TokenInfo;
 import intelligent_pay.userservice.repository.MemberRepository;
-import intelligent_pay.userservice.validator.ServiceValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
@@ -23,22 +23,20 @@ import org.springframework.transaction.annotation.Transactional;
 public class MemberCommandService {
 
     private final MemberRepository memberRepository;
-    private final ServiceValidator serviceValidator;
     private final AuthenticationManagerBuilder authenticationManagerBuilder;
     private final JwtTokenProvider jwtTokenProvider;
 
-    public String signup(MemberSignupRequest memberSignupRequest) {
-        serviceValidator.validateDuplicateEmail(memberSignupRequest.getEmail());
-
-        Member member = Member.create(memberSignupRequest);
+    public String signup(MemberSignupRequest requestDto) {
+        Member member = Member.create(requestDto);
         return memberRepository.save(member).getUsername();
     }
 
-    public TokenInfo login(MemberLoginRequest memberLoginRequest) {
-        String email = memberLoginRequest.getEmail();
-        String password = memberLoginRequest.getPassword();
+    public TokenInfo login(MemberLoginRequest requestDto) {
+        String email = requestDto.getEmail();
+        String password = requestDto.getPassword();
 
-        Member member = memberRepository.findByEmail(email);
+        Member member = memberRepository.findByEmail(email)
+                .orElseThrow(() -> new MemberCustomException(ResponseMessage.DUPLICATE_EMAIL));
         String username = member.getUsername();
 
         UsernamePasswordAuthenticationToken authenticationToken =
@@ -51,25 +49,26 @@ public class MemberCommandService {
                 .generateToken(authentication);
     }
 
-    public void updateEmail(ChangeEmailRequest changeEmailRequest, String username) {
-        String newEmail = changeEmailRequest.getEmail();
-        serviceValidator.validateDuplicateEmail(newEmail);
-
-        Member member = memberRepository.findByUsername(username);
-        member.updateEmail(newEmail);
+    public void updateEmail(ChangeEmailRequest requestDto, String username) {
+        Member member = memberRepository.findByUsername(username)
+                .orElseThrow(() -> new MemberCustomException(ResponseMessage.DUPLICATE_EMAIL));
+        member.updateEmail(requestDto.getEmail());
     }
 
-    public void updatePassword(ChangePasswordRequest changePasswordRequest, String username) {
-        serviceValidator.validatePassword(changePasswordRequest.getOldPassword(), username);
-
-        Member member = memberRepository.findByUsername(username);
-        member.updatePassword(changePasswordRequest.getNewPassword());
+    public void updatePassword(ChangePasswordRequest requestDto, String username) {
+        Member member = memberRepository.findByUsername(username)
+                .orElseThrow(() -> new MemberCustomException(ResponseMessage.DUPLICATE_EMAIL));
+        member.updatePassword(requestDto.getNewPassword(), requestDto.getOldPassword());
     }
 
-    public void withdrawByUsername(WithdrawRequest withdrawRequest, String username) {
-        serviceValidator.validatePassword(withdrawRequest.getPassword(), username);
+    public void withdrawByUsername(String username) {
+        Member member = memberRepository.findByUsername(username)
+                .orElseThrow(() -> new MemberCustomException(ResponseMessage.DUPLICATE_EMAIL));
 
-        Member member = memberRepository.findByUsername(username);
+        if (!username.equals(member.getUsername())) {
+            throw new MemberCustomException(ResponseMessage.USERNAME_NOT_MATCH);
+        }
+
         memberRepository.delete(member);
     }
 }
